moved to model/stocks.py
